Delete accumulator config & move some of the code to DapolConfig
